Made variables smaller, converted calculations to ms from s. Grinding no longer blocking Renamed grind button

+++ src/Display.h
@@ -11,7 +11,7 @@ class Display
     Display();
     void printDose1();
     void printDose2();
-    void printTime(double time);
-    void printStatistics(int numberDose1, int numberDose2);
+    void printTime(uint16_t time);
+    void printStatistics(uint16_t numberDose1, uint16_t numberDose2);
     void simpleText(char text[]);
 };
